requirements.txt
flask
pyyaml
Flask-WTF
flask-login
python-dotenv
flask-SQLAlchemy
